fix: moved useEffect hooks to screen level
